Reduce code duplication in setting up PagerAdapters
